+++ frontend/src/context/AuthContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useState } from 'react';
+import { createContext, useContext, useState, useEffect } from 'react';
 import type User from '../types/user';
 import type LoginRequest from '../types/auth';
 import * as AuthApi from '../api/auth';
@@ -6,6 +6,7 @@ import * as AuthApi from '../api/auth';
 type AuthContextType = {
     user: User | null;
     isLoggedIn: boolean;
+    isChecked: boolean;
     login: (credentials: LoginRequest) => Promise<void>;
     logout: () => Promise<void>;
 };
@@ -13,17 +14,30 @@ type AuthContextType = {
 const AuthContext = createContext<AuthContextType | null>(null);
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
-    const [user, setUser] = useState<User | null>(() => {
-        const saved = localStorage.getItem('user');
-        return saved ? JSON.parse(saved) : null;
-    });
+    const [user, setUser] = useState<User | null>(null);
+    const [isChecked, setIsChecked] = useState(false);
+
+    useEffect(() => {
+        const fetchUser = async () => {
+            AuthApi
+                .getMe()
+                .then((user) => setUser(user))
+                .catch(() => setUser(null))
+                .finally(() => setIsChecked(true));
+        };
+
+        if (user) {
+            setIsChecked(true);
+            return;
+        }
+
+        fetchUser();
+    }, []);
+
     const login = async (credentials: LoginRequest) => {
         try {
             const user: User | null = await AuthApi.loginUser(credentials);
             setUser(user);
-            if (user) {
-                localStorage.setItem('user', JSON.stringify(user));
-            }
         } catch (err) {
             console.error(err);
         }
@@ -35,12 +49,11 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
             console.error(err);
         } finally {
             setUser(null);
-            localStorage.removeItem('user');
         }
     };
 
     return (
-        <AuthContext.Provider value={{ user, isLoggedIn: !!user, login, logout }}>
+        <AuthContext.Provider value={{ user, isLoggedIn: !!user, isChecked, login, logout }}>
             {children}
         </AuthContext.Provider>
     );